+++ R/utils_api.R
@@ -6,7 +6,8 @@ fread_antares <- function(opts, file, ...) {
     response <- api_get(
       opts = opts,
       endpoint = I(file),
-      query = list(formatted = FALSE)
+      query = list(formatted = FALSE),
+      parse_result = "text"
     )
     suppressWarnings(
       tryCatch(fread(response, ...), error = function(e){